TCVP-1325 standardize project startup
