--- src/backend/TrafficCourts/Citizen.Service/Configuration/EnvironmentVariablesConfigurationSource.cs
@@ -1,9 +1,0 @@
-﻿namespace TrafficCourts.Citizen.Service.Configuration;
-
-public class EnvironmentVariablesConfigurationSource : IConfigurationSource
-{
-    public IConfigurationProvider Build(IConfigurationBuilder builder)
-    {
-        return new EnvironmentVariablesConfigurationProvider();
-    }
-}
