+++ src/backend/TrafficCourts/Citizen.Service/Instrumentation/Diagnostics.cs
@@ -1,0 +1,9 @@
+﻿using System.Diagnostics;
+
+namespace TrafficCourts.Citizen.Service
+{
+    public static class Diagnostics
+    {
+        public static readonly ActivitySource Source = new ActivitySource("citizen-api");
+    }
+}
