+++ src/backend/TrafficCourts/Citizen.Service/Services/Tickets/Search/TicketSearchService.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics;
 using System.Globalization;
-using TrafficCourts.Citizen.Service.Logging;
 using TrafficCourts.Citizen.Service.Models.Tickets;
 using TrafficCourts.Citizen.Service.Services.Tickets.Search.Common;
