+++ src/backend/TrafficCourts/Common/TrafficCourts.Common.csproj
@@ -17,6 +17,11 @@
     <PackageReference Include="Microsoft.IO.RecyclableMemoryStream" Version="2.2.0" />
     <PackageReference Include="Minio" Version="4.0.1" />
     <PackageReference Include="NodaTime" Version="3.0.10" />
+    <PackageReference Include="OpenTelemetry" Version="1.3.0-beta.1" />
+    <PackageReference Include="OpenTelemetry.Exporter.Jaeger" Version="1.3.0-beta.1" />
+    <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.0.0-rc9.3" />
+    <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.0.0-rc9.3" />
+    <PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.0.0-rc9.3" />
     <PackageReference Include="Serilog" Version="2.10.0" />
     <PackageReference Include="Serilog.AspNetCore" Version="5.0.0" />
     <PackageReference Include="Serilog.Enrichers.Context" Version="4.2.0" />
