+++ src/backend/TrafficCourts/Staff.Service/Startup.cs
@@ -1,0 +1,78 @@
+﻿using MediatR;
+using Microsoft.OpenApi.Models;
+using System.Reflection;
+using System.Text.Json.Serialization;
+using TrafficCourts.Common.Configuration;
+using TrafficCourts.Messaging;
+using TrafficCourts.Staff.Service.Authentication;
+using TrafficCourts.Staff.Service.Configuration;
+using TrafficCourts.Staff.Service.Services;
+
+namespace TrafficCourts.Staff.Service;
+
+public static class Startup
+{
+    public static void ConfigureApplication(this WebApplicationBuilder builder, Serilog.ILogger logger)
+    {
+        // Add services to the container.
+        builder.AddSerilog();
+        builder.AddOpenTelemetry(Diagnostics.Source, logger, options =>
+        {
+            options.AddSource(MassTransit.Logging.DiagnosticHeaders.DefaultListenerName);
+        });
+
+        builder.Services.AddMassTransit(builder.Configuration, logger);
+
+        // Render enums as strings rather than ints
+        builder.Services.AddControllers().AddJsonOptions(options => options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter()));
+
+        builder.Services.AddAuthentication(builder.Configuration);
+
+        // Add DisputeService
+        builder.Services.ConfigureValidatableSetting<OracleDataApiConfiguration>(builder.Configuration.GetRequiredSection(OracleDataApiConfiguration.Section));
+        builder.Services.AddSingleton<IDisputeService, DisputeService>();
+
+        builder.Services.AddMediatR(Assembly.GetExecutingAssembly());
+
+        var swagger = SwaggerConfiguration.Get(builder.Configuration);
+
+        if (swagger.Enabled)
+        {
+            logger.Information("Swagger is enabled");
+            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+            builder.Services.AddEndpointsApiExplorer();
+            builder.Services.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc("v1", new Microsoft.OpenApi.Models.OpenApiInfo
+                {
+                    Version = "v1",
+                    Title = "VTC Staff API",
+                    Description = "Violation Ticket Centre Staff API",
+                });
+
+                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
+                {
+                    Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
+                    Name = "Authorization",
+                    In = ParameterLocation.Header,
+                    Type = SecuritySchemeType.ApiKey
+                });
+
+                c.AddSecurityRequirement(new OpenApiSecurityRequirement
+        {
+            {
+                new OpenApiSecurityScheme { Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = "Bearer" } },
+                Array.Empty<string>()
+            }
+        });
+
+                c.EnableAnnotations();
+
+                // Set the comments path for the Swagger JSON and UI.
+                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+                c.IncludeXmlComments(xmlPath);
+            });
+        }
+    }
+}
