+++ src/backend/TrafficCourts/TrafficCourts.Arc.Dispute.Service/Instrumentation/Diagnostics.cs
@@ -1,0 +1,8 @@
+﻿using System.Diagnostics;
+
+namespace TrafficCourts.Arc.Dispute.Service;
+
+public static class Diagnostics
+{
+    public static readonly ActivitySource Source = new ActivitySource("arc-dispute-api");
+}
