+++ src/backend/TrafficCourts/Workflow.Service/Instrumentation/Diagnostics.cs
@@ -1,0 +1,8 @@
+﻿using System.Diagnostics;
+
+namespace TrafficCourts.Workflow.Service;
+
+public class Diagnostics
+{
+    public static readonly ActivitySource Source = new ActivitySource("workflow-service");
+}
