+++ src/backend/TrafficCourts/Workflow.Service/Program.cs
@@ -1,48 +1,18 @@
-using MassTransit;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-using TrafficCourts.Workflow.Service.Configuration;
-using TrafficCourts.Workflow.Service.Consumers;
-using TrafficCourts.Workflow.Service.Services;
-using TrafficCourts.Workflow.Service.Features.Mail;
-using TrafficCourts.Messaging;
-using Serilog;
 using TrafficCourts.Common.Configuration;
+using TrafficCourts.Common.Configuration.Validation;
+using TrafficCourts.Workflow.Service;
 
 var builder = WebApplication.CreateBuilder(args);
-var logger = GetLogger(builder);
-builder.Configuration.AddVaultSecrets(logger);
+var logger = builder.GetProgramLogger();
 
-// Add services to the container.
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-builder.Services.Configure<ArcApiConfiguration>(builder.Configuration.GetRequiredSection("ArcApiConfiguration"));
-builder.Services.Configure<OracleDataApiConfiguration>(builder.Configuration.GetRequiredSection("OracleDataApiConfiguration"));
-builder.Services.Configure<EmailConfiguration>(builder.Configuration.GetRequiredSection("EmailConfiguration"));
-builder.Services.ConfigureValidatableSetting<SmtpConfiguration>(builder.Configuration.GetRequiredSection(SmtpConfiguration.Section));
-
-builder.Services.AddTransient<IOracleDataApiService, OracleDataApiService>();
-builder.Services.AddTransient<ISubmitDisputeToArcService, SubmitDisputeToArcService>();
-builder.Services.AddTransient<ISmtpClientFactory, SmtpClientFactory>();
-builder.Services.AddTransient<IEmailSenderService, EmailSenderService>();
-
-void AddConsumers(IBusRegistrationConfigurator cfg)
-{
-    // TODO: use cfg.AddConsumers(params Type[] types) or cfg.AddConsumers(params Assembly[] assemblies)
-    cfg.AddConsumer<SubmitDisputeConsumer>();
-    cfg.AddConsumer<DisputeApprovedConsumer>();
-    cfg.AddConsumer<SendEmailConsumer>();
-}
-
-builder.Services.AddMassTransit(builder.Configuration, logger, AddConsumers);
+builder.ConfigureApplication(logger);
 
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
+var swagger = SwaggerConfiguration.Get(builder.Configuration);
+if (swagger.Enabled)
 {
     app.UseSwagger();
     app.UseSwaggerUI();
@@ -59,19 +29,18 @@ app.UseHealthChecks("/health/ready", new HealthCheckOptions()
     Predicate = (check) => check.Tags.Contains("ready")
 });
 
-app.Run();
-
-
-static Serilog.ILogger GetLogger(WebApplicationBuilder app)
+bool isDevelopment = app.Environment.IsDevelopment();
+try
 {
-    var configuration = new LoggerConfiguration()
-        .Enrich.FromLogContext()
-        .WriteTo.Console();
-
-    if (app.Environment.IsDevelopment())
-    {
-        configuration.WriteTo.Debug();
-    }
-
-    return configuration.CreateLogger();
+    app.Run();
 }
+catch (SettingsValidationException exception)
+{
+    logger.Fatal(exception, "Configuration error");
+
+    if (isDevelopment)
+    {
+        throw; // see the error in the IDE
+    }
+}
+
